# openapi/006-identity-and-authorization.yaml
openapi: 3.0.3
info:
  title: Tovu Admin API — Identity and Authorization
  version: "1.0.0"
  description: Hand-verified against the current route handlers, serializers, and identity-service validation code on 2026-08-24.
servers: [{ url: / }]
...
security: [{ adminSessionCookie: [] }, { apiKeyAuthorization: [] }]
...
paths:
  /api/admin/v1/api-keys/principals:
    post:
      operationId: create_api_key_principal
      tags: [api-keys]
      summary: Mint a grantless api_key principal
      description: >-
        Mints the ACTIVE, grantless kind='api_key' principal that issue_api_key binds to — the sole
        HTTP path that produces one. No principal_roles/principal_policies rows are written, which
        is exactly the precondition issue_api_key re-verifies, so a key's authority ends up equal to
        the policies snapshotted at issuance and never to grants the principal accumulated first.
        The workspace is the caller's own, resolved from the credential; there is no request field
        for it and therefore no cross-workspace mint. Requires apikey.manage. Handler:
        src/server/routes/admin/api-keys/create-principal.ts.
      security: [{ adminSessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateApiKeyPrincipalRequest' } } }
      responses:
        '201': { description: Created principal., content: { application/json: { schema: { $ref: '#/components/schemas/PrincipalResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/ForbiddenApiKeyManagement' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/admin/v1/auth/login:
    post:
      operationId: login
      tags: [authentication]
      summary: Create an admin session
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }
      responses:
        # Drift: runtime returns {user:{id,username}}, not LoginResponse {data:{principal,sessionExpiresAt}}.
        '200': { description: Session created, headers: { Set-Cookie: { schema: { type: string }, description: tovu_session cookie with HttpOnly, SameSite=Strict, Secure, Path=/, and Max-Age. } }, content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } } }
        # Drift: blank credentials also become this 401; the documented 400 VALIDATION_ERROR is not emitted.
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/admin/v1/auth/me:
    get:
      operationId: get_current_identity
      tags: [authentication]
      summary: Get the session or api-key principal and effective permissions
      # Drift CLOSED 2026-08-24. This previously read "runtime accepts only the tovu_session cookie;
      # the documented ApiKey Authorization header has no implementation." The runtime now accepts an
      # issued API key in an Authorization header on every /api/admin route, this one included, and
      # resolves it to the bound api_key principal with its snapshotted effectivePermissions. Both
      # the `ApiKey <raw-key>` spelling api.spec.md §2 uses and `Bearer <raw-key>` are accepted.
      security: [{ adminSessionCookie: [] }, { apiKeyAuthorization: [] }]
      responses:
        # Drift: runtime returns {user,effectivePermissions}, not {data:{principal,effectivePermissions}}.
        # For an api_key principal there is no users row, so `user.username` is the principal's
        # displayName (the value create_api_key_principal was given), not a login name.
        '200': { description: Current identity, content: { application/json: { schema: { $ref: '#/components/schemas/WhoAmIResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    adminSessionCookie: { type: apiKey, in: cookie, name: tovu_session }
    # Declared as an explicit header credential rather than `type: http, scheme: bearer` because the
    # runtime accepts TWO spellings of the same header and an http/bearer scheme could express only
    # one of them. Keeping it honest about both beats picking the tidier declaration.
    apiKeyAuthorization: { type: apiKey, in: header, name: Authorization, description: 'An issued API key, sent as `Bearer <raw-key>` or `ApiKey <raw-key>` (case-insensitive scheme). Raw keys are emitted once, by issue_api_key, and are never returned by any other endpoint.' }
...
  schemas:
    LoginRequest: { type: object, required: [username, password], properties: { username: { type: string }, password: { type: string } } }
    LoginResponse: { type: object, required: [user], properties: { user: { type: object, required: [id, username], properties: { id: { type: string }, username: { type: string } } } } }
    WhoAmIResponse: { type: object, required: [user, effectivePermissions], properties: { user: { type: object, required: [id, username], properties: { id: { type: string }, username: { type: string } } }, effectivePermissions: { type: array, items: { type: string } } } }
...
    CreateApiKeyPrincipalRequest: { type: object, required: [displayName], properties: { displayName: { type: string, minLength: 1, maxLength: 255, description: Bounds are measured after trimming; a whitespace-only value is rejected 400. Not unique-constrained — there is no login identity to collide. }, kind: { type: string, enum: [api_key], default: api_key, description: "Optional. Any other value — including 'user', 'system', and the deferred 'agent' — is rejected 400 VALIDATION_ERROR, so this endpoint can never become a second creation path for a privileged kind." } } }
    PrincipalResponse: { type: object, required: [principal], properties: { principal: { type: object, required: [id, kind, status, workspaceId, displayName], properties: { id: { type: string }, kind: { type: string, enum: [api_key] }, status: { type: string, enum: [active, disabled], description: "Always 'active' on this response — a freshly minted principal is never disabled." }, workspaceId: { type: string }, displayName: { type: string } } } } }
...
    ErrorResponse: { type: object, required: [error], properties: { error: { type: string }, code: { type: string }, details: { type: object, additionalProperties: true } } }
  responses:
    Unauthenticated: { description: Missing, invalid, expired, revoked, or disabled session; login also uses this for invalid credentials., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    ForbiddenApiKeyManagement: { description: "Caller lacks apikey.manage, OR authenticated with an API key rather than an admin session — a key may never mint, issue, or revoke another (code FORBIDDEN, details.reason='credential_kind_not_permitted'). The credential-kind check runs before the body is read and before any existence lookup, so it never reveals whether a target exists.", content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    ValidationError: { description: Invalid input or immutable/invalid target., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    RateLimited: { description: More than 10 login attempts in 60 seconds for one client IP., headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    InternalError: { description: "Internal error. Runtime body is {error: 'internal error'} with no code.", content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }